+++ src/main/java/me/junjiem/werewolf/agent/player/WerewolfPlayer.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 public class WerewolfPlayer extends AbstractPlayer {
     private final WerewolfRole role;
 
-    public WerewolfPlayer(int id, String roleName, @NonNull String service, @NonNull String apiKey, String modelName, Float temperature) {
+    public WerewolfPlayer(int id, String roleName, @NonNull String service, @NonNull String apiKey, String modelName, Double temperature) {
         super(id, roleName);
         this.role = new WerewolfRole(service, apiKey, modelName, temperature);
     }
