+++ src/main/java/me/junjiem/werewolf/agent/player/WitchPlayer.java
@@ -20,7 +20,7 @@ public class WitchPlayer extends AbstractPlayer {
     private int saveId = -1; // 昨晚救活的ID
     private int killId = -1; // 昨晚毒死的ID
 
-    public WitchPlayer(int id, String roleName, @NonNull String service, @NonNull String apiKey, String modelName, Float temperature) {
+    public WitchPlayer(int id, String roleName, @NonNull String service, @NonNull String apiKey, String modelName, Double temperature) {
         super(id, roleName);
         this.role = new WitchRole(service, apiKey, modelName, temperature);
     }
