+++ src/main/java/me/junjiem/werewolf/agent/role/HunterRole.java
@@ -25,7 +25,7 @@ public class HunterRole extends AbstractRole implements DeityRole {
     @NonNull
     private final HunterAssistant assistant;
 
-    public HunterRole(@NonNull String service, @NonNull String apiKey, String modelName, Float temperature) {
+    public HunterRole(@NonNull String service, @NonNull String apiKey, String modelName, Double temperature) {
         super(service, apiKey, modelName, temperature);
         this.assistant = AiServices.create(HunterAssistant.class, chatLanguageModel);
     }
